# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexusPay Platform API
  version: 1.0.0
  description: |
    NexusPay is a modern payment platform providing wallet, card, and transfer capabilities.

    All monetary amounts are expressed in integer minor units (kobo for NGN, cents for USD).

    **Authentication**: Most endpoints require a Bearer JWT token obtained via the `/v1/auth/login` or `/v1/auth/register` flows.

    **Idempotency**: Payment creation requests require an `Idempotency-Key` header to prevent duplicate transactions.

    **Error Format**: All errors follow a standard shape:
    ```json
    { "error": "ERROR_CODE", "message": "Human-readable message", "requestId": "optional-trace-id" }
    ```
  contact:
    name: NexusPay Support
    email: support@nexuspay.dev
    url: https://nexuspay.dev
  license:
    name: Proprietary
    url: https://nexuspay.dev/license
servers:
  - url: http://localhost:4000
    description: Local development (API Gateway)
...
paths:
...
  /v1/payments:
    post:
      operationId: createPayment
      tags: [Payments]
      summary: Initiate a payment
      description: |
        Creates a new payment intent. Requires an `Idempotency-Key` header (8-128 chars).
        Amounts are in integer minor units (e.g. 5000 = NGN 50.00).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
            example:
              amountMinor: 5000
              currency: NGN
              provider: MOCK_CARD
              metadata:
                orderId: ord_xyz789
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
              example:
                id: pay_abc123def456
                reference: NXSPAY-ABC123
                userId: usr_a1b2c3d4
                amountMinor: 5000
                currency: NGN
                provider: MOCK_CARD
                status: PENDING
                metadata:
                  orderId: ord_xyz789
                createdAt: "2026-08-25T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    get:
      operationId: listPayments
      tags: [Payments]
      summary: List payments
      description: Paginated list of payments for the authenticated user. Supports cursor-based pagination.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Number of results per page
          example: 20
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for the next page (obtained from previous response)
          example: pay_abc123def456
      responses:
        "200":
          description: Payment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentList"
              example:
                payments:
                  - id: pay_abc123def456
                    reference: NXSPAY-ABC123
                    userId: usr_a1b2c3d4
                    amountMinor: 5000
                    currency: NGN
                    provider: MOCK_CARD
                    status: COMPLETED
                    metadata: {}
                    createdAt: "2026-08-25T10:00:00Z"
                nextCursor: pay_def789ghi012
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/payments/{id}:
    get:
      operationId: getPayment
      tags: [Payments]
      summary: Get payment details
      description: Returns full payment details including ledger entries.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentDetail"
              example:
                id: pay_abc123def456
                reference: NXSPAY-ABC123
                userId: usr_a1b2c3d4
                amountMinor: 5000
                currency: NGN
                provider: MOCK_CARD
                status: COMPLETED
                providerRef: prov_ref_789
                capturedAt: "2026-08-25T10:05:00Z"
                refundedAt: null
                metadata: {}
                ledgerEntries:
                  - id: led_001
                    type: DEBIT
                    amountMinor: 5000
                    currency: NGN
                    createdAt: "2026-08-25T10:00:00Z"
                createdAt: "2026-08-25T10:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/payments/{id}/capture:
    post:
      operationId: capturePayment
      tags: [Payments]
      summary: Capture a payment
      description: Captures a previously authorized payment.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      responses:
        "200":
          description: Payment captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/payments/{id}/refund:
    post:
      operationId: refundPayment
      tags: [Payments]
      summary: Refund a payment
      description: |
        Fully refunds a captured payment. Requires the `payments:refund` or `*` scope.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      responses:
        "200":
          description: Payment refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Notifications ───────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /v1/auth/login or /v1/auth/register
    internalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
      description: Internal microservice API key for service-to-service calls

  # ── Reusable Parameters ─────────────────────────────────────────────
  parameters:
    PaymentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^pay_[a-zA-Z0-9]+$"
      description: Payment ID
      example: pay_abc123def456
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: Unique idempotency key to prevent duplicate payment creation
      example: idem_a1b2c3d4e5f6g7h8i9j0

  # ── Reusable Responses ──────────────────────────────────────────────
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: VALIDATION_ERROR
            message: "Invalid request body: email must be a valid email address"
            requestId: req_abc123
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNAUTHORIZED
            message: "Invalid or expired access token"
            requestId: req_abc123
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: FORBIDDEN
            message: "Missing required scope: payments:refund"
            requestId: req_abc123
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NOT_FOUND
            message: "Payment not found"
            requestId: req_abc123
    Conflict:
      description: Resource state conflict (e.g. already captured, duplicate key)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: CONFLICT
            message: "Payment has already been captured"
            requestId: req_abc123
    UnprocessableEntity:
      description: Request was well-formed but semantically invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNPROCESSABLE
            message: "fullName must be between 1 and 120 characters"
            requestId: req_abc123
...
  schemas:
...
    CreatePaymentRequest:
      type: object
      required: [amountMinor, provider]
      properties:
        amountMinor:
          type: integer
          exclusiveMinimum: 0
          description: Amount in minor units (kobo, cents, etc.)
          example: 5000
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          default: NGN
          description: ISO 4217 currency code (3 uppercase letters)
          example: NGN
        provider:
          type: string
          enum: [MOCK_CARD, MOCK_TRANSFER]
          description: Payment provider / method
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary key-value metadata attached to the payment
          example:
            orderId: ord_xyz789
    Payment:
      type: object
      required: [id, reference, userId, amountMinor, currency, provider, status, metadata, createdAt]
      properties:
        id:
          type: string
        reference:
          type: string
          description: Human-readable unique reference
        userId:
          type: string
        amountMinor:
          type: integer
          exclusiveMinimum: 0
        currency:
          type: string
        provider:
          type: string
          enum: [MOCK_CARD, MOCK_TRANSFER]
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, REFUNDED]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    PaymentDetail:
      allOf:
        - $ref: "#/components/schemas/Payment"
        - type: object
          properties:
            providerRef:
              type: string
              nullable: true
              description: Provider-side reference ID
            capturedAt:
              type: string
              format: date-time
              nullable: true
            refundedAt:
              type: string
              format: date-time
              nullable: true
            ledgerEntries:
              type: array
              items:
                $ref: "#/components/schemas/LedgerEntry"
              description: Double-entry ledger entries for this payment
    LedgerEntry:
      type: object
      required: [id, type, amountMinor, currency, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [DEBIT, CREDIT]
        amountMinor:
          type: integer
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
    PaymentList:
      type: object
      required: [payments, nextCursor]
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        nextCursor:
          type: string
          nullable: true
          description: Cursor for the next page; null when no more results

    # ── Notifications ───────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request body"
        requestId:
          type: string
          nullable: true
          description: Request trace ID for debugging

    # ── Users (admin) ────────────────────────────────────────────────
...